LPS-73691 Moving markAllNotificationsAsRead logic to the service
